Bug 593576 - PDF Import filter: crashes when importing large image PDF or specifying high resolution
Manually committed to master as commit adae3882d2c9955f9dd62482f9bcdb8456679701 due to code changes there.

+++ libgimp/gimplayer.c
@@ -236,7 +236,7 @@ gimp_layer_new_from_surface (gint32                image_ID,
   cairo_format_t   format;
   gint             width;
   gint             height;
-  gint             rowstride;
+  gsize            rowstride;
   gdouble          range = progress_end - progress_start;
   guint            count = 0;
   guint            done  = 0;
